Feat : 현재 금액 불러오기 로직 작성
